fix: cast non-target fp32 weights to dtype in ConvRot int8 quantizer
ConvRotInt8Quantizer.load_and_quantize left excluded/indivisible/non-target
keys (e.g. K2's "first" patch-embed Linear) at whatever dtype the checkpoint
stored, since dit_weight_dtype was never threaded through for the quantized
load path. When a checkpoint stores those weights in fp32, this breaks
F.linear with a BFloat16/Float32 dtype mismatch once the rest of the model
runs bf16 (e.g. during LoRA-wrapped sample-image inference).

ConvRotInt8Quantizer now accepts a dtype and casts any non-quantized floating
weight to it; krea2_utils.load_krea2_dit passes the real dtype through for
--convrot_int8. The --fp8_scaled path is unchanged (documented as
intentional there).

AI-assisted

+++ src/musubi_tuner/krea2/krea2_utils.py
@@ -113,7 +113,7 @@ def load_krea2_dit(
             target_keys=KREA2_FP8_OPTIMIZATION_TARGET_KEYS if fp8_scaled else None,
             exclude_keys=KREA2_FP8_OPTIMIZATION_EXCLUDE_KEYS if fp8_scaled else None,
             quantizer=(
-                ConvRotInt8Quantizer(KREA2_FP8_OPTIMIZATION_TARGET_KEYS, KREA2_FP8_OPTIMIZATION_EXCLUDE_KEYS)
+                ConvRotInt8Quantizer(KREA2_FP8_OPTIMIZATION_TARGET_KEYS, KREA2_FP8_OPTIMIZATION_EXCLUDE_KEYS, dtype=dtype)
                 if convrot_int8
                 else None
             ),
